+++ cli/dev_server.go
@@ -6,10 +6,8 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"errors"
-	"flag"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
 	"net/url"
 	"os"
@@ -27,46 +25,6 @@ import (
 	"github.com/ije/gox/utils"
 	"golang.org/x/net/html"
 )
-
-func Serve(fs *embed.FS) {
-	port := flag.Int("port", 3000, "port to serve on")
-	rootDir, _ := parseCommandFlag()
-
-	var err error
-	if rootDir == "" {
-		rootDir, err = os.Getwd()
-	} else {
-		rootDir, err = filepath.Abs(rootDir)
-		if err == nil {
-			var fi os.FileInfo
-			fi, err = os.Stat(rootDir)
-			if err == nil && !fi.IsDir() {
-				err = fmt.Errorf("stat %s: not a directory", rootDir)
-			}
-		}
-	}
-	if err != nil {
-		os.Stderr.WriteString(term.Red(err.Error()))
-		return
-	}
-
-	serv := &http.Server{
-		Addr:    fmt.Sprintf(":%d", *port),
-		Handler: &DevServer{efs: fs, rootDir: rootDir},
-	}
-	ln, err := net.Listen("tcp", serv.Addr)
-	if err != nil {
-		os.Stderr.WriteString(term.Red(err.Error()))
-		return
-	}
-
-	fmt.Printf(term.Green("Server is ready on http://localhost:%d\n"), *port)
-	err = serv.Serve(ln)
-	if err != nil {
-		os.Stderr.WriteString(term.Red(err.Error()))
-		return
-	}
-}
 
 type DevServer struct {
 	efs       *embed.FS
